add animation

+++ app/src/main/java/com/gjiazhe/layoutswitch/DensityUtil.java
@@ -6,11 +6,11 @@ import android.view.WindowManager;
 
 public class DensityUtil {
 
-    public static float dp2px(Context context, int dp) {
+    public static int dp2px(Context context, int dp) {
         DisplayMetrics dm = new DisplayMetrics();
         WindowManager windowMgr = (WindowManager) context.getSystemService(Context.WINDOW_SERVICE);
         windowMgr.getDefaultDisplay().getMetrics(dm);
-        return dm.density * dp + 0.5f;
+        return (int) (dm.density * dp + 0.5f);
 
     }
 }
